requirements.txt
fastapi
uvicorn
pyyaml
langchain
python-dotenv
motor
langchain-openai
langchain-core
langchain-community
pymongo
# lark
pydantic
urllib3
langchain-pinecone
pinecone-notebooks
pinecone-client
langchain-google-genai
# textblob
# openpyxl